requirements.txt
Flask
gunicorn
requests
flask-cors
json
random
gitpython
base64